fix event listeners

+++ js/components/Detail.js
@@ -23,7 +23,6 @@ class Detail {
 
   async render() {
     this.params = RouterUtils.getParams();
-    console.log(this.params);
     if (!this.params.id) return '';
 
     const [
@@ -47,26 +46,28 @@ class Detail {
   }
 
   async after_render() {
-    document
-      .getElementById('back_btn')
-      .addEventListener('click', async (event) => {
-        event.preventDefault();
-        window.location.replace('/#list');
-      });
-    document
-      .getElementById('edit_btn')
-      .addEventListener('click', async (event) => {
-        event.preventDefault();
-        console.log('/#edit/' + this.params.id);
-        window.location.replace('/#edit/' + this.params.id);
-      });
-    document
-      .getElementById('delete_btn')
-      .addEventListener('click', async (event) => {
-        event.preventDefault();
-        await this.dataService.deleteNote(this.params.id);
-        window.location.replace('/#list');
-      });
+    if (this.params.id) {
+      document
+        .getElementById('back_btn')
+        .addEventListener('click', async (event) => {
+          event.preventDefault();
+          window.location.replace('/#list');
+        });
+      document
+        .getElementById('edit_btn')
+        .addEventListener('click', async (event) => {
+          event.preventDefault();
+          console.log('/#edit/' + this.params.id);
+          window.location.replace('/#edit/' + this.params.id);
+        });
+      document
+        .getElementById('delete_btn')
+        .addEventListener('click', async (event) => {
+          event.preventDefault();
+          await this.dataService.deleteNote(this.params.id);
+          window.location.replace('/#list');
+        });
+    }
   }
 }
